One of these days I'll commit before pushing (RandomPoint() now takes floats rather than doubles)

+++ examples/kurt/kurt.cc
@@ -241,7 +241,7 @@ void Kurt::TimeNext(double & time_count) {
     clock_start = clock_end;
 }
 
-Point2D Kurt::RandomPoint(Point2D const &origin, double x_dist, double y_dist) {
+Point2D Kurt::RandomPoint(Point2D const &origin, float x_dist, float y_dist) {
     return { origin.x + GetRandomScalar() * x_dist, origin.y + GetRandomScalar() * y_dist };
 }
